Fix units.py to support easy swaps to different systesm.
Units.py was not nicely written and made it tricky to temporarily swap to a
different unit system.  This was needed for derived/base.py which was using
fetch_eng and thus swapping to eng units for update_archive.  Rewrote units.py
as a singleton class/module.  Derived/base.py now importing plain fetch but
swapping units as needed.

Other:
- Move structured_array into Ska.Numpy.
- Various non-impacting cleanup.