Fix regression in update_archive.py in stats H5 creation
Original update_stats() method did not add initial data when
first creating stats H5 files.  This was changed so initial
ingest also put in stats data, but for some reason this
caused a problem with a record at the end of the first ingest.
Backed this out to the original logic.