Improve support for units so fetch{,_eng,_sci} don't clobber each other
Previously there was just a single module .units that stored the
current state of units.  It was easy to inadvertently clobber the
current units if another module imported a different set.  The updated
units.py module implements a Units class that behaves similarly to
the old module but is sandboxed in a class instance.