Fix MSID/MSIDset interpolate to only include values in fetched range
After this change the interpolate() function will only output data
values that are within the fetched time ranges for the input MSID
or MSIDs.  In the case of an MSIDset this is the smallest intersection
of all set MSID time ranges.

Previously there was a questionable behavior where the original query
range was used to determine the range and values were extrapolated
(nearest-neighbor) to fill in gaps at the edges.  This was especially
a problem for the common case of doing a fetch query with no stop
value so it went up to NOW.  In this case new data were invented
from the end of the archive up to now.