Support more generalized unit conversions
Previously the unit conversion support was from CXC to ENG or SCI.
Now it goes in any direction, but in particular from ENG to SCI
or CXC.

In addition the conversion transforms are completely specified
and if a conversion is requested outside of those known transforms
then a warning is issued.  It should probably be an exception but
don't want to break code immediately.